clean up or repo and add readme for submission
