+++ finviz/searchstocks/dash_apps/finished_apps/searchstocks.py
@@ -11,6 +11,15 @@ from dash.dependencies import Output
 import plotly.express as px
 import plotly.graph_objs as go
 from django_plotly_dash import DjangoDash
+from searchstocks.views import read_ticker_symbols
+from searchstocks.views import get_symbols
+from searchstocks.views import get_sectors
+from searchstocks.views import dropdown_values
+from searchstocks.views import colormap_value
+from searchstocks.views import stock_information
+from searchstocks.views import get_company_stats
+
+
 
 ##############################################################################################
 #Parameters and instantiating classes
@@ -18,136 +27,77 @@ obj_iexcloud = iexCloud()
 external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
 app = DjangoDash('searchstocks', external_stylesheets=external_stylesheets)
 
-def read_ticker_symbols():
-    try:
-        df_tickers = pd.read_sql('SELECT * FROM db_ingestion_tickers;', engine_string)
-        df_tickerstats = pd.read_sql('SELECT * FROM db_ingestion_tickerstats', engine_string)
-        df = pd.merge(df_tickers, df_tickerstats, how='inner', on='Symbol')
-        
-        # #Format Market Capitalisation
-        # df['Market Capitalization Ranges'] = 'Placeholder'
-        # df.loc[(df['Market Capitalization'] <= 250000000), 'Market Capitalization Ranges'] = '$0 - $250Million'
-        # df.loc[(df['Market Capitalization'] >= 250000000) & (df['Market Capitalization'] <= 500000000), 'Market Capitalization Ranges'] = '$250Million - $500Million'
-        # df.loc[(df['Market Capitalization'] >= 500000000) & (df['Market Capitalization'] <= 1000000000), 'Market Capitalization Ranges'] = '$500Million - $1Billion'
-        # df.loc[(df['Market Capitalization'] >= 1000000000) & (df['Market Capitalization'] <= 10000000000), 'Market Capitalization Ranges'] = '$1Billion - $10Billion'
-        # df.loc[(df['Market Capitalization'] >= 1000000000) & (df['Market Capitalization'] <= 50000000000), 'Market Capitalization Ranges'] = '$10Billion - $50Billion'
-        # df.loc[(df['Market Capitalization'] >= 5000000000) & (df['Market Capitalization'] <= 100000000000), 'Market Capitalization Ranges'] = '$50Billion - $100Billion'
-        # df.loc[(df['Market Capitalization'] >= 100000000000) & (df['Market Capitalization'] <= 500000000000), 'Market Capitalization Ranges'] = '$100Billion - $500Billion'
-        # df.loc[(df['Market Capitalization'] >= 500000000000), 'Market Capitalization Ranges'] = 'Greater than $500Billion'
+# def read_ticker_symbols():
+#     try:
+#         df_tickers = pd.read_sql('SELECT * FROM db_ingestion_tickers;', engine_string)
+#         df_tickerstats = pd.read_sql('SELECT * FROM db_ingestion_tickerstats', engine_string)
+#         df = pd.merge(df_tickers, df_tickerstats, how='inner', on='Symbol')
+#         return df
+#     except Exception as e:
+#         print(str(e))
 
-        # #Format Dividend
-        # df['Dividend Ranges'] = 'Placeholder'
-        # df.loc[(df['Dividend'] <= 0), 'Dividend Ranges'] = '0%'
-        # df.loc[(df['Dividend'] >= 0) & (df['Dividend'] <= 0.01), 'Dividend Ranges'] = '0% - 1%'
-        # df.loc[(df['Dividend'] >= 0.01) & (df['Dividend'] <= 0.025), 'Dividend Ranges'] = '1% - 2.5%'
-        # df.loc[(df['Dividend'] >= 0.025) & (df['Dividend'] <= 0.05), 'Dividend Ranges'] = '2.5% - 5%'
-        # df.loc[(df['Dividend'] >= 0.05) & (df['Dividend'] <= 0.075), 'Dividend Ranges'] = '5% - 7.5%'
-        # df.loc[(df['Dividend'] >= 0.075) & (df['Dividend'] <= 0.1), 'Dividend Ranges'] = '7.5% - 10%'
-        # df.loc[(df['Dividend'] >= 0.1) & (df['Dividend'] <= 0.15), 'Dividend Ranges'] = '10% - 15%'
-        # df.loc[(df['Dividend'] >= 0.15) & (df['Dividend'] <= 0.2), 'Dividend Ranges'] = '15% - 20%'
-        # df.loc[(df['Dividend'] >= 2), 'Dividend Ranges'] = 'Greater than 20%'
+# def get_symbols():
+#     dropdown_options = []
+#     df = read_ticker_symbols()
+#     for index, row in df.iterrows():
+#         dropdown_options.append({'label': row['Name'], 'value': row['Symbol']})
+#     return dropdown_options
 
-        # #Format Dividend
-        # df['PE Ratio Ranges'] = 'Placeholder'
-        # df.loc[(df['PE Ratio'] <= -10), 'PE Ratio Ranges'] = 'Less than -10'
-        # df.loc[(df['PE Ratio'] >= -10) & (df['PE Ratio'] <= 0), 'PE Ratio Ranges'] = '-10 - 0'
-        # df.loc[(df['PE Ratio'] >= 0) & (df['PE Ratio'] <= 10), 'PE Ratio Ranges'] = '0 - 10'
-        # df.loc[(df['PE Ratio'] >= 10) & (df['PE Ratio'] <= 20), 'PE Ratio Ranges'] = '10 - 20'
-        # df.loc[(df['PE Ratio'] >= 20) & (df['PE Ratio'] <= 30), 'PE Ratio Ranges'] = '20 - 30'
-        # df.loc[(df['PE Ratio'] >= 30) & (df['PE Ratio'] <= 40), 'PE Ratio Ranges'] = '30 - 40'
-        # df.loc[(df['PE Ratio'] >= 40) & (df['PE Ratio'] <= 50), 'PE Ratio Ranges'] = '40 - 50'
-        # df.loc[(df['PE Ratio'] >= 50) & (df['PE Ratio'] <= 75), 'PE Ratio Ranges'] = '50 - 75'
-        # df.loc[(df['PE Ratio'] >= 75) & (df['PE Ratio'] <= 100), 'PE Ratio Ranges'] = '75 - 100'
-        # df.loc[(df['PE Ratio'] >= 100), 'PE Ratio Ranges'] = 'Greater than 100'
+# def get_sectors():
+#     dropdown_options = []
+#     df = read_ticker_symbols()
+#     sector_lst = list(df['Sector'].unique())
+#     for i in sector_lst:
+#         dropdown_options.append({'label':i, 'value':i})
+#     return dropdown_options
 
-        # #Format Revenue
-        # df['Revenue Ranges'] = 'Placeholder'
-        # df.loc[(df['Revenue'] <= 250000000), 'Revenue Ranges'] = '$0 - $250Million'
-        # df.loc[(df['Revenue'] >= 250000000) & (df['Revenue'] <= 500000000), 'Revenue Ranges'] = '$250Million - $500Million'
-        # df.loc[(df['Revenue'] >= 500000000) & (df['Revenue'] <= 1000000000), 'Revenue Ranges'] = '$500Millionn- $1Billion'
-        # df.loc[(df['Revenue'] >= 1000000000) & (df['Revenue'] <= 10000000000), 'Revenue Ranges'] = '$1Billion - $10Billion'
-        # df.loc[(df['Revenue'] >= 10000000000) & (df['Revenue'] <= 50000000000), 'Revenue Ranges'] = '$10Billion - $50Billion'
-        # df.loc[(df['Revenue'] >= 50000000000) & (df['Revenue'] <= 100000000000), 'Revenue Ranges'] = '$50Billion - $100Billion'
-        # df.loc[(df['Revenue'] >= 100000000000), 'Revenue Ranges'] = 'Greater than $100Billion'
-        # #243,198,000
+# def dropdown_values(column_name):
+#     dropdown_options = []
+#     df = read_ticker_symbols()
+#     lst = list(df[column_name].unique())
+#     for i in lst:
+#         dropdown_options.append({'label':i, 'value':i})
+#     return dropdown_options
 
-        # #Format EBITDA
-        # df['EBITDA Ranges'] = 'Placeholder'
-        # df.loc[(df['EBITDA'] < -100000000), 'EBITDA Ranges'] = 'Less than -$100Million'
-        # df.loc[(df['EBITDA'] < -100000000) & (df['EBITDA'] <= -50000000), 'EBITDA Ranges'] = '-$100Million- -$50Million'
-        # df.loc[(df['EBITDA'] >= -50000000) & (df['EBITDA'] <= 0), 'EBITDA Ranges'] = '-$50Million - $0'
-        # df.loc[(df['EBITDA'] >= 0) & (df['EBITDA'] <= 250000000), 'EBITDA Ranges'] = '$0 - $250Million'
-        # df.loc[(df['EBITDA'] >= 250000000) & (df['EBITDA'] <= 500000000), 'EBITDA Ranges'] = '$250Million - $500Million'
-        # df.loc[(df['EBITDA'] >= 500000000) & (df['EBITDA'] <= 1000000000), 'EBITDA Ranges'] = '$500Million - $1Billion'
-        # df.loc[(df['EBITDA'] >= 1000000000) & (df['EBITDA'] <= 10000000000), 'EBITDA Ranges'] = '$1Billion - $10Billion'
-        # df.loc[(df['EBITDA'] >= 10000000000), 'EBITDA Ranges'] = 'Greater than $10Billion'
+# def colormap_value(table_name, range):
+#     try:
+#         df_colormaps = pd.read_sql(f'SELECT * FROM db_ingestion_{table_name}_colormaps;', engine_string)
+#         df_colormaps = df_colormaps[df_colormaps['Metric'] == range]
+#         print(df_colormaps)
 
-        return df
-    except Exception as e:
-        print(str(e))
+#         color = df_colormaps['Color'].iloc[0]
+#         return color
+#     except Exception as e:
+#         print(str(e))
 
-def get_symbols():
-    dropdown_options = []
-    df = read_ticker_symbols()
-    for index, row in df.iterrows():
-        dropdown_options.append({'label': row['Name'], 'value': row['Symbol']})
-    return dropdown_options
+# def stock_information(ticker, metric):
+#     df_tickerstats = pd.read_sql(f'SELECT * FROM db_ingestion_tickerstats', engine_string)
+#     df_tickerstats = df_tickerstats[df_tickerstats['Symbol'] == ticker]
 
-def get_sectors():
-    dropdown_options = []
-    df = read_ticker_symbols()
-    sector_lst = list(df['Sector'].unique())
-    for i in sector_lst:
-        dropdown_options.append({'label':i, 'value':i})
-    return dropdown_options
-
-def dropdown_values(column_name):
-    dropdown_options = []
-    df = read_ticker_symbols()
-    lst = list(df[column_name].unique())
-    for i in lst:
-        dropdown_options.append({'label':i, 'value':i})
-    return dropdown_options
-
-def colormap_value(table_name, range):
-    try:
-        df_colormaps = pd.read_sql(f'SELECT * FROM db_ingestion_{table_name}_colormaps;', engine_string)
-        df_colormaps = df_colormaps[df_colormaps['Metric'] == range]
-        print(df_colormaps)
-
-        color = df_colormaps['Color'].iloc[0]
-        return color
-    except Exception as e:
-        print(str(e))
-
-def stock_information(ticker, metric):
-    df_tickerstats = pd.read_sql(f'SELECT * FROM db_ingestion_tickerstats', engine_string)
-    df_tickerstats = df_tickerstats[df_tickerstats['Symbol'] == ticker]
-
-    if metric == 'marketcap':
-        updated_metric = 'Market Capitalization Ranges'
-    if metric == 'dividend':
-        updated_metric = 'Dividend Ranges'
-    if metric == 'ebitda':
-        updated_metric = 'EBITDA Ranges'
-    if metric == 'peratio':
-        updated_metric = 'PE Ratio Ranges'
-    if metric == 'revenue':
-        updated_metric = 'Revenue Ranges'
+#     if metric == 'marketcap':
+#         updated_metric = 'Market Capitalization Ranges'
+#     if metric == 'dividend':
+#         updated_metric = 'Dividend Ranges'
+#     if metric == 'ebitda':
+#         updated_metric = 'EBITDA Ranges'
+#     if metric == 'peratio':
+#         updated_metric = 'PE Ratio Ranges'
+#     if metric == 'revenue':
+#         updated_metric = 'Revenue Ranges'
 
 
-    range = df_tickerstats[updated_metric].iloc[0]
+#     range = df_tickerstats[updated_metric].iloc[0]
 
-    return range
+#     return range
 
 
-def get_company_stats():
-    try:
-        df = pd.read_sql(f'SELECT * FROM db_ingestion_tickerstats', engine_string)
-    except Exception as e:
-        print(str(e))
+# def get_company_stats():
+#     try:
+#         df = pd.read_sql(f'SELECT * FROM db_ingestion_tickerstats', engine_string)
+#     except Exception as e:
+#         print(str(e))
 
-    return df
+#     return df
 
 
 
@@ -315,8 +265,7 @@ app.layout = html.Div([
 #Callbacks
 @app.callback(
     Output("time-series-chart", "figure"), 
-    [Input("button", 'n_clicks'),
-    Input("stockselector", "value"),
+    [Input("stockselector", "value"),
     Input("sectorselector", "value"),
     Input("marketcap_selector", "value"),
     Input("dividend_selector", "value"),
@@ -326,83 +275,79 @@ app.layout = html.Div([
     Input("colormap", "value")]
     )
 
-def time_series_stock(n_clicks, ticker_dropdown, sector_dropdown, marketcap_selector, 
-                        dividend_selector, pe_selector, revenue_selector,
-                        ebitda_selector, colormap):
+def time_series_stock(ticker_dropdown, sector_dropdown, marketcap_selector, 
+                    dividend_selector, pe_selector, revenue_selector,
+                    ebitda_selector, colormap):
 
     graphs = []
 
 
-    if n_clicks > 0:
+    #Flatten list
+    if any(isinstance(i, list) for i in ticker_dropdown):
+        ticker_dropdown = [item for elem in ticker_dropdown for item in elem]
 
+    #graphs = []
 
-        #Flatten list
-        if any(isinstance(i, list) for i in ticker_dropdown):
-            ticker_dropdown = [item for elem in ticker_dropdown for item in elem]
+    #Draw time series of a single stock
+    for ticker in ticker_dropdown:
+        stock_df = obj_iexcloud.get_max_time_series_df(ticker)
+        graphs.append(go.Scatter(
+            x = stock_df['Date'],
+            y = stock_df[ticker],
+            mode = 'lines',
+            name = ticker,
+            textposition = 'bottom center',
+        ))
 
-        #graphs = []
+    parameters = {
+        'Market Capitalization Ranges': marketcap_selector,
+        'Dividend Ranges': dividend_selector,
+        'PE Ratio Ranges': pe_selector,
+        'Revenue Ranges': revenue_selector,
+        'EBITDA Ranges': ebitda_selector
+    }
+    filled_parameters = {}
+    for key, value in parameters.items():
+        if value:
+            filled_parameters[key] = value
 
-        #Draw time series of a single stock
-        for ticker in ticker_dropdown:
-            stock_df = obj_iexcloud.get_max_time_series_df(ticker)
+    #Draw time series of sectors and filter
+    sector_df = read_ticker_symbols()
+    sector_df = sector_df[sector_df['Sector'].isin(sector_dropdown)]
+    print(sector_df)
+    #if marketcap_dropdown:
+    for key in  filled_parameters:
+        #sector_df = sector_df.loc[(sector_df['Market Capitalization Ranges'].isin(marketcap_dropdown))]
+        sector_df = sector_df.loc[(sector_df[key].isin(filled_parameters[key]))]
+    stock_list = list(sector_df['Symbol'])
+    for stock in stock_list:
+        stock_df = obj_iexcloud.get_max_time_series_df(stock)
+        #Choose color based on metric
+        if colormap == 'none':
             graphs.append(go.Scatter(
-                x = stock_df['Date'],
-                y = stock_df[ticker],
-                mode = 'lines',
-                name = ticker,
-                textposition = 'bottom center',
+            x = stock_df['Date'],
+            y = stock_df[stock],
+            mode = 'lines',
+            name = stock,
+            textposition = 'bottom center',
+            ))
+        else:
+            metric = colormap
+            range = stock_information(stock, metric)
+            color = colormap_value(metric, range)
+
+            graphs.append(go.Scatter(
+            x = stock_df['Date'],
+            y = stock_df[stock],
+            mode = 'lines',
+            name = stock,
+            textposition = 'bottom center',
+            line=dict(color=f'#{color}'),
             ))
 
-        parameters = {
-            'Market Capitalization Ranges': marketcap_selector,
-            'Dividend Ranges': dividend_selector,
-            'PE Ratio Ranges': pe_selector,
-            'Revenue Ranges': revenue_selector,
-            'EBITDA Ranges': ebitda_selector
-        }
-        filled_parameters = {}
-        for key, value in parameters.items():
-            if value:
-                filled_parameters[key] = value
-
-        #Draw time series of sectors and filter
-        sector_df = read_ticker_symbols()
-        sector_df = sector_df[sector_df['Sector'].isin(sector_dropdown)]
-        print(sector_df)
-        #if marketcap_dropdown:
-        for key in  filled_parameters:
-            #sector_df = sector_df.loc[(sector_df['Market Capitalization Ranges'].isin(marketcap_dropdown))]
-            sector_df = sector_df.loc[(sector_df[key].isin(filled_parameters[key]))]
-        stock_list = list(sector_df['Symbol'])
-        for stock in stock_list:
-            stock_df = obj_iexcloud.get_max_time_series_df(stock)
-            #Choose color based on metric
-            if colormap == 'none':
-                graphs.append(go.Scatter(
-                x = stock_df['Date'],
-                y = stock_df[stock],
-                mode = 'lines',
-                name = stock,
-                textposition = 'bottom center',
-                ))
-            else:
-                metric = colormap
-                range = stock_information(stock, metric)
-                color = colormap_value(metric, range)
-
-                graphs.append(go.Scatter(
-                x = stock_df['Date'],
-                y = stock_df[stock],
-                mode = 'lines',
-                name = stock,
-                textposition = 'bottom center',
-                line=dict(color=f'#{color}'),
-                ))
-
-        if any(isinstance(i, list) for i in graphs):
-            graphs = [item for elem in graphs for item in elem]
-        
-        n_clicks = 0
+    if any(isinstance(i, list) for i in graphs):
+        graphs = [item for elem in graphs for item in elem]
+    
 
     fig = {
             'data': graphs,
