+++ requirements.txt
@@ -1,9 +1,150 @@
-django
-django_plotly_dash
-channels
-channels_redis
-pandas
-requests
-dash
-dash_bootstrap_components
-SQLAlchemy
+aiohttp==3.8.1
+aioredis==1.3.1
+aiosignal==1.2.0
+appdirs==1.4.4
+appnope==0.1.3
+asgiref==3.5.0
+asttokens==2.0.5
+async-timeout==4.0.2
+attrs==21.4.0
+autobahn==22.1.1
+Automat==20.2.0
+backcall==0.2.0
+backports.zoneinfo==0.2.1
+boto3==1.24.1
+botocore==1.27.1
+Brotli==1.0.9
+cached-property==1.5.2
+certifi==2021.10.8
+cffi==1.15.0
+channels==3.0.4
+channels-redis==3.3.1
+chardet==4.0.0
+charset-normalizer==2.0.11
+click==8.0.3
+clikit==0.6.2
+cmdstanpy==0.9.68
+constantly==15.1.0
+convertdate==2.4.0
+crashtest==0.3.1
+cryptography==36.0.1
+cycler==0.11.0
+Cython==0.29.28
+daphne==3.0.2
+dash==2.1.0
+dash-bootstrap-components==1.0.3
+dash-core-components==2.0.0
+dash-html-components==2.0.0
+dash-renderer==1.9.1
+dash-table==5.0.0
+datapackage==1.15.2
+debugpy==1.6.0
+decorator==5.1.1
+Django==4.0.2
+django-plotly-dash==1.6.6
+dpd-components==0.1.0
+entrypoints==0.4
+ephem==4.1.3
+et-xmlfile==1.1.0
+executing==0.8.3
+Flask==2.0.2
+Flask-Compress==1.10.1
+fonttools==4.33.3
+frozenlist==1.3.0
+greenlet==1.1.2
+hijri-converter==2.2.3
+hiredis==2.0.0
+holidays==0.13
+httpstan==4.7.2
+hyperlink==21.0.0
+idna==3.3
+ijson==3.1.4
+importlib-resources==5.7.1
+incremental==21.3.0
+ipykernel==6.13.0
+ipython==8.3.0
+isodate==0.6.1
+itsdangerous==2.0.1
+jedi==0.18.1
+Jinja2==3.0.3
+jmespath==1.0.0
+joblib==1.1.0
+jsonlines==3.0.0
+jsonpointer==2.3
+jsonschema==4.6.0
+jupyter-client==7.3.1
+jupyter-core==4.10.0
+kiwisolver==1.4.2
+korean-lunar-calendar==0.2.1
+linear-tsv==1.1.0
+LunarCalendar==0.0.9
+MarkupSafe==2.0.1
+marshmallow==3.15.0
+matplotlib==3.5.2
+matplotlib-inline==0.1.3
+msgpack==1.0.3
+multidict==6.0.2
+nest-asyncio==1.5.5
+nltk==3.7
+numpy==1.22.4
+openpyxl==3.0.10
+packaging==21.3
+pandas==1.4.2
+parso==0.8.3
+pastel==0.2.1
+pexpect==4.8.0
+pickleshare==0.7.5
+Pillow==9.1.0
+plotly==5.5.0
+prompt-toolkit==3.0.29
+prophet==1.0.1
+psutil==5.9.0
+ptyprocess==0.7.0
+pure-eval==0.2.2
+pyasn1==0.4.8
+pyasn1-modules==0.2.8
+pycparser==2.21
+Pygments==2.12.0
+pylev==1.4.0
+PyMeeus==0.5.11
+pyOpenSSL==22.0.0
+pyparsing==3.0.9
+pyrsistent==0.18.1
+pysimdjson==3.2.0
+pystan==2.19.1.1
+python-dateutil==2.8.2
+pytz==2021.3
+pyzmq==22.3.0
+regex==2022.4.24
+requests==2.27.1
+rfc3986==2.0.0
+s3transfer==0.6.0
+scikit-learn==1.0.2
+scipy==1.8.1
+service-identity==21.1.0
+setuptools-git==1.2
+six==1.16.0
+sklearn==0.0
+SQLAlchemy==1.4.31
+sqlparse==0.4.2
+stack-data==0.2.0
+tableschema==1.20.2
+tabulator==1.53.5
+tenacity==8.0.1
+threadpoolctl==3.1.0
+tornado==6.1
+tqdm==4.64.0
+traitlets==5.2.0
+Twisted==22.1.0
+txaio==21.2.1
+typing-extensions==4.0.1
+ujson==5.2.0
+unicodecsv==0.14.1
+urllib3==1.26.8
+wcwidth==0.2.5
+webargs==8.1.0
+Werkzeug==2.0.2
+xlrd==2.0.1
+yarl==1.7.2
+zipp==3.8.0
+zope.interface==5.4.0
